WIP: RFC: pin entrpoint/commands on fixed core
For ultra-low-latency workloads, it is helpful to guarantee
that the container entry point, and any other foreign process
run in the container (kubectl exec...) are always scheduled
on fixed and predictable cpu in the container cpuset (e.g. not in a
random one).

This patch implements this optional behaviour depending on
container annotations.

Signed-off-by: Francesco Romani <fromani@redhat.com>